refactor: update type definitions and fix some build error on osx

+++ src/core/include/platform/mac/mac_platform.hpp
@@ -8,12 +8,12 @@ namespace atlas
 {
 struct MacPlatformTypes
 {
-    typedef unsigned long long      uint64;
+    typedef unsigned long           uint64;
     typedef unsigned int            uint32;
     typedef unsigned short int      uint16;
     typedef unsigned char           uint8;
 
-    typedef signed long long        int64;
+    typedef signed long             int64;
     typedef signed int              int32;
     typedef signed short int        int16;
     typedef signed char             int8;
